[RISCV] Allow shrink wrapping for RISC-V
Enabling shrink wrapping requires ensuring the insertion point of the
epilogue is correct for MBBs without a terminator, in which case the
instruction to adjust the stack pointer is the last instruction in the
block.

To ensure that this change is safe when used with the save/restore
optimisation, the prologue/epilogue insertion points need to be checked.
A libcall save requires a temporary register, and a libcall restore
requires the ability to insert a tail call.